fix for build dev

+++ build_dev.ps1
@@ -13,8 +13,11 @@ Invoke-ps2exe .\WarframeMarketTool.ps1 .\build_dev\WarframeMarketTool.exe
 # Copy the Views folder to the dist directory
 Copy-Item -Path ".\Views" -Destination ".\build_dev\Views" -Recurse -Force
 
-# Copy the config.json file to the dist directory
-Copy-Item -Path ".\config_base.json" -Destination ".\build_dev\config.json" -Force
+# Copy the config.json file to the dist directory if not exist
+if (-not (Test-Path -Path ".\build_dev\config.json")) {
+    Copy-Item -Path ".\config_base.json" -Destination ".\build_dev\config.json" -Force
+}
+
 
 Write-Host "###################################################################################"
 Write-Host "Build DEV process completed."
